feat: Updated room page to allow user to select from their linked platforms
In the room page, users will be able to add songs. First a user selects from their linked platforms so the app knows which platforms to search within.

App first retrieves linked platforms (currently kept track of in local memory) and only shows those platforms as options.

Also updated sample search results to show platform on black background (this version of the platform is not clickable as if it were it would attempt to either be set as an option for search retrieval (described above) or begin user authentication (done on separate page))