fix: Able to add tracks to queue and play them via media player
Some bugs in rendering queue items.

Clicking a song in the queue will skip to that song (but not remove it yet) and play it in the player. Can pause / play current song. When a song is currently playing and the user clicks another song in the queue, the new song metadata will transfer to the media player but the song will not be heard.

To be fixed: Queue rendering issue, queued song to play while a song is currently playing (after being overwritten, of course)