chore: Programatic redirect URL for soundcloud
Soundcloud only allows for one redirectURL so when going from using local host or where the application is hosted on vercel as development options, a new push was made to alter the url.

Now can just detect if local host or not, and the only change required is to change the redirect URL on soundcloud's site.