feat: Added development toggle
As API quotas are limited for fetching data, implemented a toggle so that I can set the testing mode as "Development".

With this on/off toggle, I can then choose to get test data (i.e. YouTube iframe, or other platform data) from a file (loaded into when making an actual call first).

When off: searches will search via API calls as expected.
When on: searches will not search but pull pre-searched data from a local file.

Note: This will not make it to full production, but just for testing.