feat: Added logic to update access tokens when expired
Wish to later update access tokens when they are close to expiring (10 minutes before) instead of exactly when they do

As auth for youtube is setup as a client side web app, auth process does not return a refresh token. So, to refresh the its access token requires additional authorization requests.